# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GCRF Library Management System API
  version: 1.0.0
  description: |
    RESTful API documentation for GCRF (国创睿峰) Intelligent Library Management System.

    This system provides comprehensive library management capabilities including:
    - Authentication and authorization
    - Book catalog management
    - Reader management
    - Circulation (borrow, return, renew)
    - System administration

    **Architecture**: Microservices with Spring Cloud Gateway

    **Base URL**: All requests go through the API Gateway at `http://localhost:8080`

    **Authentication**: JWT Bearer token required for most endpoints (see Authentication section)

  contact:
    name: GCRF Library Team
    email: support@gcrf-library.com
  license:
    name: Proprietary
    url: https://gcrf-library.com/license
servers:
  - url: http://localhost:8080
    description: Development server (Local)
  - url: https://api.gcrf-library.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/readers:
    get:
      tags: [Readers]
      summary: List readers with pagination
      description: Query reader list with filters
      parameters:
        - name: pageNum
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
        - name: keyword
          in: query
          description: Search keyword (name, readerId, phone)
          schema:
            type: string
        - name: readerType
          in: query
          description: Reader type filter
          schema:
            type: string
            enum: [STUDENT, TEACHER, STAFF, PUBLIC]
        - name: cardStatus
          in: query
          description: Card status filter
          schema:
            type: string
            enum: [PENDING, ACTIVE, SUSPENDED, CANCELLED]
      responses:
        '200':
          description: Reader list retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedReaderResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags: [Readers]
      summary: Create reader
      description: Register a new library reader and generate library card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReaderCreateRequest'
      responses:
        '200':
          description: Reader created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReaderDetailVO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/readers/{id}:
    get:
      tags: [Readers]
      summary: Get reader details
      description: Retrieve detailed information of a specific reader
      parameters:
        - name: id
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reader details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReaderDetailVO'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags: [Readers]
      summary: Update reader
      description: Update reader information
      parameters:
        - name: id
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReaderUpdateRequest'
      responses:
        '200':
          description: Reader updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReaderDetailVO'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags: [Readers]
      summary: Delete reader
      description: Soft delete reader (logic delete)
      parameters:
        - name: id
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reader deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/readers/readerId/{readerId}:
    get:
      tags: [Readers]
      summary: Get reader by card number
      description: Retrieve reader information by reader card number
      parameters:
        - name: readerId
          in: path
          required: true
          description: Reader card number
          schema:
            type: string
      responses:
        '200':
          description: Reader details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReaderDetailVO'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/readers/{id}/activate:
    post:
      tags: [Readers]
      summary: Activate library card
      description: Activate reader's library card to enable borrowing
      parameters:
        - name: id
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card activated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReaderDetailVO'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/readers/{id}/suspend:
    post:
      tags: [Readers]
      summary: Suspend library card
      description: Suspend reader's library card (report lost)
      parameters:
        - name: id
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card suspended successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReaderDetailVO'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/readers/{id}/cancel:
    post:
      tags: [Readers]
      summary: Cancel library card
      description: Cancel reader's library card permanently
      parameters:
        - name: id
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card cancelled successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ReaderDetailVO'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ========================================
  # Circulation Management APIs
  # ========================================
...
  /api/v1/circulation/reader/{readerId}:
    get:
      tags: [Circulation]
      summary: Get reader's circulation records
      description: Query circulation history for a specific reader
      parameters:
        - name: readerId
          in: path
          required: true
          description: Reader ID
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          description: Record status filter (0=Borrowed, 1=Returned, 2=Overdue)
          schema:
            type: integer
            enum: [0, 1, 2]
      responses:
        '200':
          description: Circulation records retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CirculationRecord'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ========================================
  # Role Management APIs
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header:
        ```
        Authorization: Bearer <your-jwt-token>
        ```
...
  schemas:
...
    ReaderCreateRequest:
      type: object
      required:
        - readerId
        - name
        - readerType
      properties:
        readerId:
          type: string
          maxLength: 50
          description: Reader card number (unique)
          example: "R202400001"
        name:
          type: string
          maxLength: 100
          example: "王小明"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          example: "MALE"
        idCard:
          type: string
          pattern: '^[1-9]\d{5}(18|19|20)\d{2}((0[1-9])|(1[0-2]))(([0-2][1-9])|10|20|30|31)\d{3}[0-9Xx]$'
          example: "110101199001011234"
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
          example: "13912345678"
        email:
          type: string
          format: email
          example: "wangxiaoming@example.com"
        readerType:
          type: string
          enum: [STUDENT, TEACHER, STAFF, PUBLIC]
          example: "STUDENT"
        department:
          type: string
          maxLength: 200
          example: "计算机科学与技术学院"
        majorOrPosition:
          type: string
          maxLength: 200
          example: "软件工程"
        depositAmount:
          type: integer
          minimum: 0
          description: Deposit amount in cents (分)
          example: 10000
        maxBorrowQuantity:
          type: integer
          minimum: 1
          maximum: 100
          example: 5
        remarks:
          type: string
          maxLength: 1000
    ReaderUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        email:
          type: string
          format: email
        department:
          type: string
          maxLength: 200
        majorOrPosition:
          type: string
          maxLength: 200
        maxBorrowQuantity:
          type: integer
          minimum: 1
          maximum: 100
        remarks:
          type: string
          maxLength: 1000
    ReaderDetailVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        readerId:
          type: string
          example: "R202400001"
        name:
          type: string
          example: "王小明"
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER]
          example: "MALE"
        idCard:
          type: string
          example: "110101199001011234"
        phone:
          type: string
          example: "13912345678"
        email:
          type: string
          example: "wangxiaoming@example.com"
        readerType:
          type: string
          enum: [STUDENT, TEACHER, STAFF, PUBLIC]
          example: "STUDENT"
        department:
          type: string
          example: "计算机科学与技术学院"
        majorOrPosition:
          type: string
          example: "软件工程"
        cardStatus:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED, CANCELLED]
          example: "ACTIVE"
        depositAmount:
          type: integer
          example: 10000
        currentBorrowedCount:
          type: integer
          example: 3
        maxBorrowQuantity:
          type: integer
          example: 5
        totalBorrowedCount:
          type: integer
          example: 25
        overdueCount:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CirculationRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        readerId:
          type: integer
          format: int64
          example: 1
        bookId:
          type: integer
          format: int64
          example: 10
        borrowDate:
          type: string
          format: date-time
          example: "2024-10-28T14:30:00"
        dueDate:
          type: string
          format: date-time
          example: "2024-11-27T14:30:00"
        returnDate:
          type: string
          format: date-time
          nullable: true
          example: null
        renewCount:
          type: integer
          example: 0
        status:
          type: integer
          description: "0=Borrowed, 1=Returned, 2=Overdue"
          example: 0
        overdueDays:
          type: integer
          example: 0
        fine:
          type: integer
          description: Fine amount in cents (分)
          example: 0
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          description: Response code (200 for success)
          example: 200
        message:
          type: string
          description: Response message
          example: "操作成功"
        data:
          description: Response data (varies by endpoint)
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
          example: 1698765432000
        traceId:
          type: string
          description: Request trace ID for debugging
          example: "trace-abc123xyz"
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: Error code
          example: 400
        message:
          type: string
          description: Error message
          example: "参数验证失败"
        data:
          nullable: true
        timestamp:
          type: integer
          format: int64
          example: 1698765432000
        traceId:
          type: string
          example: "trace-abc123xyz"
    PageResult:
      type: object
      properties:
        records:
          type: array
          items: {}
          description: List of records
        total:
          type: integer
          format: int64
          description: Total number of records
          example: 100
        pageNum:
          type: integer
          description: Current page number (1-based)
          example: 1
        pageSize:
          type: integer
          description: Page size
          example: 10
        pages:
          type: integer
          description: Total number of pages
          example: 10
...
    PagedReaderResponse:
      allOf:
        - $ref: '#/components/schemas/PageResult'
        - type: object
          properties:
            records:
              type: array
              items:
                $ref: '#/components/schemas/ReaderDetailVO'
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing token
              value:
                code: 401
                message: "未提供认证令牌"
                data: null
                timestamp: 1698765432000
            invalidToken:
              summary: Invalid token
              value:
                code: 401
                message: "令牌无效或已过期"
                data: null
                timestamp: 1698765432000
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: 404
                message: "请求的资源不存在"
                data: null
                timestamp: 1698765432000
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              value:
                code: 400
                message: "参数验证失败：用户名不能为空"
                data: null
                timestamp: 1698765432000
...